Consumer driven PACT tests prototype
