+++ build.gradle
@@ -167,6 +167,11 @@ configure(springProjects) {
       mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
     }
   }
+
+  test {
+    useJUnitPlatform()
+    systemProperties['pact.rootDir'] = "../pact-tests"
+  }
 }
 
 wrapper {
