+++ customers/build.gradle
@@ -1,3 +1,19 @@
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath 'au.com.dius:pact-jvm-provider-gradle_2.12:3.6.2'
+    }
+}
+
+plugins {
+    id "jacoco"
+    id "au.com.dius.pact" version "3.6.2"
+}
+
+version = '0.0.1'
+
 dependencies {
     providedRuntime("org.springframework.boot:spring-boot-starter-tomcat")
     compile("org.springframework.cloud:spring-cloud-starter-netflix-eureka-client")
@@ -6,12 +22,35 @@ dependencies {
     testCompile("org.axonframework:axon-test:${axonVersion}")
 }
 
-version = '0.0.1'
-
-apply plugin: "jacoco"
-
 jacocoTestReport {
     reports {
         html.destination file("${buildDir}/jacocoHtml")
     }
+}
+
+pact {
+
+    serviceProviders {
+
+        // You can define as many as you need, but each must have a unique name
+        Customers {
+            // All the provider properties are optional, and have sensible defaults (shown below)
+            protocol = 'http'
+            host = 'localhost'
+            port = 8082
+            path = '/'
+
+            // Or if you have many pact files in a directory
+            hasPactsWith('manyConsumers') {
+
+                // Will define a consumer for each pact file in the directory.
+                // Consumer name is read from contents of pact file
+                pactFileLocation = file('../pact-tests')
+
+            }
+
+        }
+
+    }
+
 }
