+++ customers/src/test/java/com/ultimatesoftware/banking/customer/unit/CustomerControllerUnitTests.java
@@ -4,6 +4,7 @@ import com.ultimatesoftware.banking.customer.domain.models.Customer;
 import com.ultimatesoftware.banking.customer.service.controllers.CustomerController;
 import com.ultimatesoftware.banking.customer.service.repositories.CustomerRepository;
 
+import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.runner.RunWith;
@@ -11,6 +12,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.springframework.data.domain.Example;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -23,6 +25,7 @@ import static org.mockito.Mockito.when;
 public class CustomerControllerUnitTests {
     @InjectMocks
     private CustomerController customerController;
+
     @Mock
     private CustomerRepository customerRepository;
 
@@ -44,7 +47,7 @@ public class CustomerControllerUnitTests {
         customerController.getCustomer(id);
 
         //assert
-        verify(customerRepository, times(1)).findOne(id);
+        verify(customerRepository, times(1)).findOne(Example.of(customer));
     }
 
     @Test
@@ -71,13 +74,13 @@ public class CustomerControllerUnitTests {
         customerController.updateCustomer(id, customer);
 
         //assert
-        verify(customerRepository, times(1)).findOne(id);
+        verify(customerRepository, times(1)).findOne(Example.of(customer));
     }
 
     @Test
     public void onUpdateCustomersWithExistingId_repositorySaveIsCalled() {
         //arrange
-        when(customerRepository.findOne(id)).thenReturn(customer);
+        when(customerRepository.findOne(Example.of(customer))).thenReturn(Optional.of(customer));
 
         //act
         customerController.updateCustomer(id, customer);
@@ -101,19 +104,19 @@ public class CustomerControllerUnitTests {
         customerController.deleteCustomers(id);
 
         //assert
-        verify(customerRepository, times(1)).findOne(id);
+        verify(customerRepository, times(1)).findOne(Example.of(customer));
     }
 
     @Test
     public void onDeleteCustomersWithExistingId_repositoryDeleteIsCalled() {
         //arrange
-        when(customerRepository.findOne(id)).thenReturn(customer);
+        when(customerRepository.findOne(Example.of(customer))).thenReturn(Optional.of(customer));
 
         //act
         customerController.deleteCustomers(id);
 
         //assert
-        verify(customerRepository, times(1)).delete(id);
+        verify(customerRepository, times(1)).delete(customer);
     }
 
     @Test
