+++ transactions/build.gradle
@@ -10,7 +10,7 @@ dependencies {
     compile("org.axonframework:axon-spring-boot-starter:${axonVersion}")
     compile group: 'au.com.dius', name: 'pact-jvm-consumer-junit5_2.12', version: '3.6.2'
     testCompile("org.axonframework:axon-test:${axonVersion}")
-    testCompile 'io.rest-assured:rest-assured:3.1.1'
+    testCompile 'io.rest-assured:rest-assured:3.3.0'
 }
 
 version = '0.0.1'
