--- transactions/src/test/java/com/ultimatesoftware/banking/transactions/contract/CustomerConsumerContractTest.java
@@ -1,26 +1,0 @@
-package com.ultimatesoftware.banking.transactions.contract;
-
-import au.com.dius.pact.consumer.Pact;
-import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
-import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
-import au.com.dius.pact.consumer.junit5.PactTestFor;
-import au.com.dius.pact.model.RequestResponsePact;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-@ExtendWith(PactConsumerTestExt.class)
-@PactTestFor(providerName = "ArticlesProvider", port = "1234")
-public class CustomerConsumerContractTest {
-
-    @Pact(provider="test_provider", consumer="test_consumer")
-    public RequestResponsePact createPact(PactDslWithProvider builder) {
-        return builder
-            .given("test state")
-            .uponReceiving("ExampleJavaConsumerPactTest test interaction")
-            .path("/")
-            .method("GET")
-            .willRespondWith()
-            .status(200)
-            .body("{\"responsetest\": true}")
-            .toPact();
-    }
-}
